Use big_int for telegram id

+++ data/models/user.py
@@ -1,10 +1,10 @@
 from data.models.base_model import BaseModel
-from peewee import AutoField, TextField, IntegerField
+from peewee import AutoField, TextField, BigIntegerField
 
 
 class User(BaseModel):
     id = AutoField(column_name='id')
-    telegram_id = IntegerField(column_name='telegram_id')
+    telegram_id = BigIntegerField(column_name='telegram_id')
     telegram_name = TextField(column_name='telegram_name')
 
     class Meta:
